Make the modal component more accessible
Update the App component so that it passes a modal ref to handle and force focus when the modal should be displayed. Update the Modal to handle it, and added A11y attributes to the rendered Modal martkup. Also add a close button to the modal.

Add additional tests for the modal to test it for accessibility.

TODO: Add more to handle keyboard presses